+++ skbio/stats/tests/test_power.py
@@ -21,19 +21,37 @@ from skbio.stats.power import (subsample_power,
                                _calculate_power,
                                _compare_distributions,
                                _calculate_power_curve,
+                               _check_subsample_power_inputs,
                                bootstrap_power_curve,
-                               paired_subsamples)
+                               paired_subsamples
+                               )
 
 
 class PowerAnalysisTest(TestCase):
 
     def setUp(self):
-        # Defines a testing function
+
+        # Defines a testing functions
         def test_meta(ids, meta, cat, div):
             """Checks thhe div metric with a kruskal wallis"""
             out = [meta.loc[id_, div] for id_ in ids]
             return kruskal(*out)[1]
+
+        def meta_f(x):
+            """Applies `test_meta` to a result"""
+            return test_meta(x, self.meta, 'INT', 'DIV')
+
+        def f(x):
+            """returns the p value of a kruskal wallis test"""
+            return kruskal(*x)[1]
+
+        def multi_f(x):
+            pass
+
         self.test_meta = test_meta
+        self.f = f
+        self.meta_f = meta_f
+
         # Sets the random seed
         np.random.seed(5)
         # Sets up the distributions of data for use
@@ -46,8 +64,6 @@ class PowerAnalysisTest(TestCase):
         self.alpha = np.power(10, np.array([-1, -1.301, -2, -3])).round(3)
         # Sets up a vector of samples
         self.num_samps = np.arange(10, 100, 10)
-        # Sets up the test function, a rank-sum test
-        self.f = lambda x: kruskal(*x)[1]
         # Sets up a mapping file
         meta = {'GW': {'INT': 'N', 'ABX': np.nan, 'DIV': 19.5, 'AGE': '30s',
                        'SEX': 'M'},
@@ -74,7 +90,6 @@ class PowerAnalysisTest(TestCase):
                 'NR': {'INT': 'Y', 'ABX': 'Y', 'DIV': 15.7, 'AGE': '20s',
                        'SEX': 'F'}}
         self.meta = pd.DataFrame.from_dict(meta, orient='index')
-        self.meta_f = lambda x: test_meta(x, self.meta, 'INT', 'DIV')
         self.counts = np.array([5, 15, 25, 35, 45])
         self.powers = [np.array([[0.105, 0.137, 0.174, 0.208, 0.280],
                                  [0.115, 0.135, 0.196, 0.204, 0.281],
@@ -99,33 +114,13 @@ class PowerAnalysisTest(TestCase):
         self.cat = "AGE"
         self.control_cats = ['INT', 'ABX']
 
-    def test_subsample_power_matched_relationship_error(self):
-        with self.assertRaises(ValueError):
-            subsample_power(self.f,
-                            samples=[np.ones((2)), np.ones((5))],
-                            draw_mode="matched")
-
-    def test_subsample_power_min_observations_error(self):
-        with self.assertRaises(ValueError):
-            subsample_power(self.f,
-                            samples=[np.ones((2)), np.ones((5))])
-
-    def test_subsample_power_interval_error(self):
-        with self.assertRaises(ValueError):
-            subsample_power(self.f,
-                            samples=[np.ones((3)), np.ones((5))],
-                            min_observations=2,
-                            min_counts=5,
-                            counts_interval=1000,
-                            max_counts=7)
-
     def test_subsample_power_defaults(self):
         test_p, test_c = subsample_power(self.f, self.pop,
                                          num_iter=10, num_runs=5)
         self.assertEqual(test_p.shape, (5, 4))
         npt.assert_array_equal(np.array([10, 20, 30, 40]), test_c)
 
-    def test_subsample_power(self):
+    def test_subsample_power_counts(self):
         test_p, test_c = subsample_power(self.f,
                                          samples=self.pop,
                                          num_iter=10,
@@ -134,23 +129,22 @@ class PowerAnalysisTest(TestCase):
         self.assertEqual(test_p.shape, (2, 5))
         npt.assert_array_equal(np.arange(5, 50, 10), test_c)
 
-    def test_subsample_paired_power_min_observations_error(self):
-        with self.assertRaises(ValueError):
-            subsample_paired_power(self.f,
-                                   self.meta,
-                                   cat=self.cat,
-                                   control_cats=self.control_cats)
+    def test_subsample_power_matches(self):
+        test_p, test_c = subsample_power(self.f,
+                                         samples=self.pop,
+                                         num_iter=10,
+                                         num_runs=5,
+                                         draw_mode="matched")
+        self.assertEqual(test_p.shape, (5, 4))
+        npt.assert_array_equal(np.array([10, 20, 30, 40]), test_c)
 
-    def test_subsample_paired_power_interval_error(self):
-        with self.assertRaises(ValueError):
-            subsample_paired_power(self.f,
-                                   self.meta,
-                                   cat='INT',
-                                   control_cats=['SEX', 'AGE'],
-                                   min_observations=2,
-                                   counts_interval=12,
-                                   min_counts=5,
-                                   max_counts=7)
+    def test_subsample_power_multi_p(self):
+        test_p, test_c = subsample_power(lambda x: np.array([0.5, 0.5]),
+                                         samples=self.pop,
+                                         num_iter=10,
+                                         num_runs=5)
+        self.assertEqual(test_p.shape, (5, 4, 2))
+        npt.assert_array_equal(np.array([10, 20, 30, 40]), test_c)
 
     def test_subsample_paired_power(self):
         known_c = np.array([1, 2, 3, 4, 5])
@@ -167,8 +161,24 @@ class PowerAnalysisTest(TestCase):
                                                 num_iter=10,
                                                 num_runs=2)
         # Test the output shapes are sane
-        npt.assert_array_equal(test_p.shape, (2, 5))
+        self.assertEqual(test_p.shape, (2, 5))
         npt.assert_array_equal(known_c, test_c)
+
+    def test_subsample_paired_power_multi_p(self):
+        def f(x):
+            return np.array([0.5, 0.5, 0.005])
+        cat = 'INT'
+        control_cats = ['SEX']
+        # Tests for the control cats
+        test_p, test_c = subsample_paired_power(f,
+                                                meta=self.meta,
+                                                cat=cat,
+                                                control_cats=control_cats,
+                                                min_observations=1,
+                                                counts_interval=1,
+                                                num_iter=10,
+                                                num_runs=2)
+        self.assertEqual(test_p.shape, (2, 5, 3))
 
     def test__check_strs_str(self):
         self.assertTrue(_check_strs('string'))
@@ -235,28 +245,35 @@ class PowerAnalysisTest(TestCase):
         # Checks the test value
         npt.assert_almost_equal(known, test)
 
+    def test__calculate_power_n(self):
+        crit = 0.025
+        known = np.array([0.5, 0.5])
+        alpha = np.vstack((self.alpha, self.alpha))
+        test = _calculate_power(alpha, crit)
+        npt.assert_almost_equal(known, test)
+
     def test__compare_distributions_mode_error(self):
         with self.assertRaises(ValueError):
-            _compare_distributions(self.f, self.samps, mode='fig')
+            _compare_distributions(self.f, self.samps, 1, mode='fig')
 
     def test__compare_distributions_count_error(self):
         with self.assertRaises(ValueError):
-            _compare_distributions(self.f, self.samps, counts=[1, 2, 3],
+            _compare_distributions(self.f, self.samps, 1, counts=[1, 2, 3],
                                    num_iter=100)
 
     def test__compare_distributions_matched_length_error(self):
         with self.assertRaises(ValueError):
-            _compare_distributions(self.f, [np.ones((5)), np.zeros((6))],
+            _compare_distributions(self.f, [np.ones((5)), np.zeros((6))], 1,
                                    mode="matched")
 
     def test__compare_distributions_sample_counts_error(self):
         with self.assertRaises(ValueError):
-            _compare_distributions(self.f, [self.pop[0][:5], self.pop[1]],
-                                   25)
+            _compare_distributions(self.f, [self.pop[0][:5], self.pop[1]], 1,
+                                   counts=25)
 
     def test__compare_distributions_all_mode(self):
         known = np.ones((100))*0.0026998
-        test = _compare_distributions(self.f, self.samps, num_iter=100)
+        test = _compare_distributions(self.f, self.samps, 1, num_iter=100)
         npt.assert_allclose(known, test, 5)
 
     def test__compare_distributions_matched_mode(self):
@@ -264,13 +281,77 @@ class PowerAnalysisTest(TestCase):
         known_mean = 0.162195
         known_std = 0.121887
         known_shape = (100,)
-        # Sets the sample value
         # Tests the sample value
-        test = _compare_distributions(self.f, self.pop, mode='matched',
+        test = _compare_distributions(self.f, self.pop, 1, mode='matched',
                                       num_iter=100)
         npt.assert_allclose(known_mean, test.mean(), rtol=0.1, atol=0.02)
         npt.assert_allclose(known_std, test.std(), rtol=0.1, atol=0.02)
         self.assertEqual(known_shape, test.shape)
+
+    def test__compare_distributions_multiple_returns(self):
+        known = np.array([[1, 1, 1], [2, 2, 2], [3, 3, 3]])
+
+        def f(x):
+            return np.array([1, 2, 3])
+
+        test = _compare_distributions(f, self.pop, 3, mode='matched',
+                                      num_iter=3)
+        npt.assert_array_equal(known, test)
+
+    def test_check_subsample_power_inputs_min_observations(self):
+        with self.assertRaises(ValueError):
+            _check_subsample_power_inputs(self.f,
+                                          samples=[np.ones((2)), np.ones((2))],
+                                          min_observations=3)
+
+    def test_check_subsample_power_inputs_matched_mode(self):
+        with self.assertRaises(ValueError):
+            _check_subsample_power_inputs(self.f,
+                                          samples=[np.ones((2)), np.ones((5))],
+                                          draw_mode="matched",
+                                          min_observations=1)
+
+    def test_check_subsample_power_inputs_counts(self):
+        with self.assertRaises(ValueError):
+            _check_subsample_power_inputs(self.f,
+                                          samples=[np.ones((3)), np.ones((5))],
+                                          min_observations=2,
+                                          min_counts=5,
+                                          counts_interval=1000,
+                                          max_counts=7)
+
+    def test_check_subsample_power_inputs_ratio(self):
+        with self.assertRaises(ValueError):
+            _check_subsample_power_inputs(self.f,
+                                          self.samps,
+                                          min_observations=1,
+                                          ratio=np.array([1, 2, 3]))
+
+    def test_check_subsample_power_inputs_test(self):
+        # Defines a test function
+        def test(x):
+            return 'Hello World!'
+        with self.assertRaises(TypeError):
+            _check_subsample_power_inputs(test,
+                                          self.samps,
+                                          min_observations=1)
+
+    def test_check_sample_power_inputs(self):
+        # Defines the know returns
+        known_num_p = 1
+        known_ratio = np.ones((2))
+        known_counts = np.arange(2, 10, 2)
+        # Runs the code for the returns
+        test_ratio, test_num_p, test_counts = \
+            _check_subsample_power_inputs(self.f,
+                                          self.samps,
+                                          min_observations=1,
+                                          counts_interval=2,
+                                          max_counts=10)
+        # Checks the returns are sane
+        self.assertEqual(known_num_p, test_num_p)
+        npt.assert_array_equal(known_ratio, test_ratio)
+        npt.assert_array_equal(known_counts, test_counts)
 
     def test__calculate_power_curve_ratio_error(self):
         with self.assertRaises(ValueError):
@@ -279,17 +360,16 @@ class PowerAnalysisTest(TestCase):
                                    num_iter=100)
 
     def test__calculate_power_curve_default(self):
-        # Sets the know output
+        # Sets the known output
         known = np.array([0.509, 0.822, 0.962, 0.997, 1.000, 1.000, 1.000,
-                          1.000,  1.000])
-
-        # Generates the test values.
+                          1.000, 1.000])
+        # Generates the test values
         test = _calculate_power_curve(self.f,
                                       self.pop,
                                       self.num_samps,
                                       num_iter=100)
         # Checks the samples returned sanely
-        npt.assert_allclose(test, known, rtol=0.1, atol=0.1)
+        npt.assert_allclose(test, known, rtol=0.1, atol=0.01)
 
     def test__calculate_power_curve_alpha(self):
         # Sets the know output
@@ -327,11 +407,13 @@ class PowerAnalysisTest(TestCase):
                                1.000, 1.000,  1.000])
         known_bound = np.array([0.03, 0.02, 0.01, 0.01, 0.00, 0.00, 0.00, 0.00,
                                 0.00])
+
         # Generates the test values
         test_mean, test_bound = bootstrap_power_curve(self.f,
                                                       self.pop,
                                                       self.num_samps,
                                                       num_iter=100)
+
         # Checks the function returned sanely
         npt.assert_allclose(test_mean, known_mean, rtol=0.05, atol=0.05)
         npt.assert_allclose(test_bound, known_bound, rtol=0.1, atol=0.01)